Inplement basics of CSV Parser and Linked List
This is the initial commit which includes the following:

1. Linked list
The linked list implementation can:
    - Add a new node to the end of a linked list chain
    - Delete a specified node from a linked list chain
    - Search a linked list chain for a node containing a specific
      sensor name
    - Count how many nodes are in a linked list chain

2. CSV Parser
The CSV parser implementation can:
    - Parse a given .csv file (included as input.csv) and generate
      a linked list with the values contained in the file
    - Identify when all sensors with the same timestamp have been
      added and pass the linked list on to the fused sensor
      algorithm
    - Stop parsing at the end of the file

3. General
A "high limit" (-h) and a "low limit" (-l) can be passed in by the
user to specify the range of acceptable sensor values. Nodes that
are outside of that range will not be passed onto the algorithm.