Add Doxygen automated documentation
This adds the markup code required to have Doxygen recognize and
create documentation based on the comments in the code. In order to
generate the documentation, you will need to do the following (this
assumes you have Doxygen installed).

 - In the sensorfusion directory, run: doxygen sensor_fusion_config
 - HTML documentation will be generated in sensorfusion/doc/html

Git will ignore everything in this directory, so we should add the
documentation generation to the makefile