Fix issue with time conversion
There was a floating point error in how I was converting the time from
the CSV file which lead to some of the times being off by a minute.
Instead of relying on floating point math, I've figured out how to use
strtok_r to parse multiple strings at once and am just parsing the time
and storing it directly instead.